requirements.txt
Package                  Version   
------------------------ ----------
cachetools               3.1.0     
certifi                  2018.11.29
chardet                  3.0.4     
google-api-core          1.7.0     
google-auth              1.6.2     
google-cloud-vision      0.35.2    
googleapis-common-protos 1.5.6     
grpcio                   1.18.0    
idna                     2.8       
instagram-scraper        1.5.41    
pip                      19.0.1    
protobuf                 3.6.1     
pyasn1                   0.4.5     
pyasn1-modules           0.2.4     
pytz                     2018.9    
requests                 2.21.0    
rsa                      4.0       
setuptools               40.8.0    
six                      1.12.0    
tqdm                     4.30.0    
urllib3                  1.24.1    
wheel                    0.32.3    
